Fix some instances of -Wshadow.
Bug: 794619
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3141022
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#919333}

+++ chrome/browser/safe_browsing/android/safe_browsing_bridge.cc
@@ -61,7 +61,8 @@ static jboolean JNI_SafeBrowsingBridge_GetSafeBrowsingExtendedReportingManaged(
 }
 
 static jint JNI_SafeBrowsingBridge_GetSafeBrowsingState(JNIEnv* env) {
-  return safe_browsing::GetSafeBrowsingState(*GetPrefService());
+  return static_cast<jint>(
+      safe_browsing::GetSafeBrowsingState(*GetPrefService()));
 }
 
 static void JNI_SafeBrowsingBridge_SetSafeBrowsingState(JNIEnv* env,
